requirements.txt
Flask==2.3.3
gunicorn==21.2.0
supabase==2.0.3
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
httpx==0.24.1
python-dotenv==1.0.0
redis==5.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
psycopg2-binary==2.9.9
postgrest>=0.10.0
python-dateutil==2.8.2
pytesseract==0.3.10
urllib3==2.0.7
pytz==2023.3
Flask-Limiter==3.5.0
pytest==7.4.3
pytest-mock==3.12.0
pdf2image==1.17.0
Pillow==10.3.0
schedule==1.2.1
PyYAML==6.0.1
reportlab==4.0.9
PyPDF2==3.0.1
opencv-python==4.9.0.80
